+++ CHANGELOG.md
@@ -1,33 +1,87 @@
-# [1.0.0-develop.3](https://github.com/sebbo2002/js-template-test/compare/v1.0.0-develop.2...v1.0.0-develop.3) (2021-05-08)
+## [1.1.1-develop.12](https://github.com/sebbo2002/fitness-first-ical-proxy/compare/v1.1.1-develop.11...v1.1.1-develop.12) (2022-05-13)
 
+## [1.1.1-develop.11](https://github.com/sebbo2002/fitness-first-ical-proxy/compare/v1.1.1-develop.10...v1.1.1-develop.11) (2022-05-09)
 
-### Bug Fixes
+## [1.1.1-develop.10](https://github.com/sebbo2002/fitness-first-ical-proxy/compare/v1.1.1-develop.9...v1.1.1-develop.10) (2022-05-05)
 
-* **CI:** Fix npm github release ([8355e8b](https://github.com/sebbo2002/js-template-test/commit/8355e8b3658b534ad359d03147d6b6d559de3340))
+## [1.1.1-develop.9](https://github.com/sebbo2002/fitness-first-ical-proxy/compare/v1.1.1-develop.8...v1.1.1-develop.9) (2022-05-05)
 
-# [1.0.0-develop.2](https://github.com/sebbo2002/js-template-test/compare/v1.0.0-develop.1...v1.0.0-develop.2) (2021-05-08)
+## [1.1.1-develop.8](https://github.com/sebbo2002/fitness-first-ical-proxy/compare/v1.1.1-develop.7...v1.1.1-develop.8) (2022-05-02)
 
+## [1.1.1-develop.7](https://github.com/sebbo2002/fitness-first-ical-proxy/compare/v1.1.1-develop.6...v1.1.1-develop.7) (2022-05-01)
 
-### Bug Fixes
+## [1.1.1-develop.6](https://github.com/sebbo2002/fitness-first-ical-proxy/compare/v1.1.1-develop.5...v1.1.1-develop.6) (2022-05-01)
 
-* Docker Tags ([7d8dcd5](https://github.com/sebbo2002/js-template-test/commit/7d8dcd579f3d796e6a07d77e7da8700eb44a4ce0))
+## [1.1.1-develop.5](https://github.com/sebbo2002/fitness-first-ical-proxy/compare/v1.1.1-develop.4...v1.1.1-develop.5) (2022-04-28)
 
-# 1.0.0-develop.1 (2021-05-08)
+## [1.1.1-develop.4](https://github.com/sebbo2002/fitness-first-ical-proxy/compare/v1.1.1-develop.3...v1.1.1-develop.4) (2022-04-22)
 
+## [1.1.1-develop.3](https://github.com/sebbo2002/fitness-first-ical-proxy/compare/v1.1.1-develop.2...v1.1.1-develop.3) (2022-04-18)
 
-### Bug Fixes
+## [1.1.1-develop.2](https://github.com/sebbo2002/fitness-first-ical-proxy/compare/v1.1.1-develop.1...v1.1.1-develop.2) (2022-04-09)
 
-* Cannot read property 'env' of undefined ([7f64b7a](https://github.com/sebbo2002/js-template-test/commit/7f64b7a9c3b8f42dd3f608be0251fa1fdaad32d8))
-* **CI:** Fix docker build ([7da7438](https://github.com/sebbo2002/js-template-test/commit/7da74380621eed1134a87d0c01d4da8b45e58525))
+## [1.1.1-develop.1](https://github.com/sebbo2002/fitness-first-ical-proxy/compare/v1.1.0...v1.1.1-develop.1) (2022-03-31)
 
-
-### Features
-
-* test ([2008ee5](https://github.com/sebbo2002/js-template-test/commit/2008ee563dab2203ac6c51d55e06631af86b58b2))
-
-# 1.0.0-develop.1 (2021-05-07)
+# [1.1.0](https://github.com/sebbo2002/fitness-first-ical-proxy/compare/v1.0.1...v1.1.0) (2022-03-07)
 
 
 ### Features
 
-* test ([2008ee5](https://github.com/sebbo2002/js-template-test/commit/2008ee563dab2203ac6c51d55e06631af86b58b2))
+* Add "Kinderschwimmen" to Fitness First Gendarmenmarkt ([58462c3](https://github.com/sebbo2002/fitness-first-ical-proxy/commit/58462c34fb003a00a6690a294b5b75a0c1f4ee41))
+
+# [1.1.0-develop.1](https://github.com/sebbo2002/fitness-first-ical-proxy/compare/v1.0.1...v1.1.0-develop.1) (2022-02-28)
+
+
+### Features
+
+* Add "Kinderschwimmen" to Fitness First Gendarmenmarkt ([58462c3](https://github.com/sebbo2002/fitness-first-ical-proxy/commit/58462c34fb003a00a6690a294b5b75a0c1f4ee41))
+
+## [1.0.1](https://github.com/sebbo2002/fitness-first-ical-proxy/compare/v1.0.0...v1.0.1) (2021-12-13)
+
+
+### Bug Fixes
+
+* **CI:** Fix DockerHub container release ([01b7534](https://github.com/sebbo2002/fitness-first-ical-proxy/commit/01b753406d1f1ef24a949c7d7b946d99b779d013))
+
+# 1.0.0 (2021-11-10)
+
+### chore
+
+* Remove node.js 10 Support ([2b910c0](https://github.com/sebbo2002/fitness-first-ical-proxy/commit/2b910c09bc8a41085fc4472159494d8738d5521e))
+
+
+### Features
+
+* first commit ([bcca5cc](https://github.com/sebbo2002/fitness-first-ical-proxy/commit/bcca5cc0dc3d58b3ebee9fa2a5a9f60c4dd0e474))
+
+
+### BREAKING CHANGES
+
+* Removed support for node.js v10
+
+# [1.0.0-develop.3](https://github.com/sebbo2002/fitness-first-ical-proxy/compare/v1.0.0-develop.2...v1.0.0-develop.3) (2021-12-05)
+
+
+### Bug Fixes
+
+* **CI:** Fix DockerHub container release ([01b7534](https://github.com/sebbo2002/fitness-first-ical-proxy/commit/01b753406d1f1ef24a949c7d7b946d99b779d013))
+
+
+# [1.0.0-develop.2](https://github.com/sebbo2002/fitness-first-ical-proxy/compare/v1.0.0-develop.1...v1.0.0-develop.2) (2021-11-05)
+
+# 1.0.0-develop.1 (2021-10-22)
+
+
+### chore
+
+* Remove node.js 10 Support ([2b910c0](https://github.com/sebbo2002/fitness-first-ical-proxy/commit/2b910c09bc8a41085fc4472159494d8738d5521e))
+
+
+### Features
+
+* first commit ([bcca5cc](https://github.com/sebbo2002/fitness-first-ical-proxy/commit/bcca5cc0dc3d58b3ebee9fa2a5a9f60c4dd0e474))
+
+
+### BREAKING CHANGES
+
+* Removed support for node.js v10